Computer-Generated-Hologram/requirements.txt
numpy
pillow
matplotlib
scipy
opencv-python
tqdm